% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayCount.R
\name{assayCount}
\alias{assayCount}
\title{The number of assays in the projects}
\usage{
assayCount(projectAccession = c("PXD000001", "PXD008959"),
  accessionRownames = TRUE)
}
\arguments{
\item{projectAccession}{character, a project accession number,
for example: PXD000001.}
}
\value{
The number of assays in the projects.
}
\description{
The number of assays in the projects
}
\examples{
{
\dontrun{
# Search for one project accession
assayCount("PXD000001")

# Search for two project accessions
assayCount(c("PXD000001", "PXD008959"))
}
}
}
