% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayDetail.R
\name{assayDetail}
\alias{assayDetail}
\title{The details of assays returned by searching}
\usage{
assayDetail(accession = c("22134", "22135"), accessionRownames = TRUE)
}
\arguments{
\item{accession}{character, a vector of assay accession. The default is
c("22134", "22135")}
}
\value{
a data.frame of assay details, including:
projectAccession, assayAccession, title, shortLabel, species,
sampleDetails, ptmNames, instrumentNames, keywords, experimentalFactor,
proteinCount, peptideCount, uniquePeptideCount, identifiedSpectrumCount,
totalSpectrumCount, ms2Annotation, chromatogram, softwares, diseases,
quantMethods, and contacts.
}
\description{
The details of assays returned by searching
}
\examples{
{
\dontrun{
# Search for one accession
assayDetail("22134")

# Search for two accessions
assayDetail(c("22134", "22135"))
}
}
}
