% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assayList.R
\name{assayList}
\alias{assayList}
\title{The assay entries information in the projects}
\usage{
assayList(projectAccession, accessionRownames = TRUE)
}
\arguments{
\item{projectAccession}{character, a project accession number
for example: PXD000001.}
}
\value{
a data.frame of the details of assay entries information, including:
projectAccession (string, optional): the project this assay belongs to,
assayAccession (string, optional): the accession assigned to the assay,
title (string, optional): the title give to the assay by the submitter,
shortLabel (string, optional): the short label give to the assay by the submitter,
species (array[string], optional): the species reported for this assay,
sampleDetails (array[string], optional): sample details reported for this assay
ptmNames (array[string], optional): reported modifications,
instrumentNames (array[string], optional): reported instrument information,
keywords (string, optional): additional keywords added to this assay,
experimentalFactor (string, optional): experimental factors reported for this assay,
proteinCount (integer, optional): number of proteins in this assay,
peptideCount (integer, optional): number of peptides in this assay,
uniquePeptideCount (integer, optional): number of unique peptides in this assay,
identifiedSpectrumCount (integer, optional): number of identified spectra in this assay,
totalSpectrumCount (integer, optional): total number of spectra in this assay,
ms2Annotation (boolean, optional): flag if ms level 2 annotation is available,
chromatogram (boolean, optional): flag to indicate if a chromatogram is available,
softwares (array[string], optional): software used for the data/result generation,
diseases (array[string], optional): disease annotation provided for this assay (if applicable),
quantMethods (array[string], optional): quantification methods used,
contacts (array[ContactDetail], optional): contact persons, usually the submitter of the dataset,
}
\description{
The assay entries information in the projects
}
\examples{
{
\dontrun{
# Search for one project accession
assayList("PXD000001")

# Search for two project accessions
assayList(c("PXD000001", "PXD008959"))
}
}
}
