% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCountAssay.R
\name{fileCountAssay}
\alias{fileCountAssay}
\title{The number of files in the assay}
\usage{
fileCountAssay(assayAccession = c("22134", "22135"),
  accessionRownames = TRUE)
}
\arguments{
\item{assayAccession}{character, a vector of project accessions.
The default is c("22134", "22135")}
}
\value{
a data.frame, the number of files in the assay.
}
\description{
The number of files in the assay
}
\examples{
{
\dontrun{
fileCountAssay("22135")
fileCountAssay(c("22134", "22135"))
}
}
}
