% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCountProject.R
\name{fileCountProject}
\alias{fileCountProject}
\title{The number of files in the projects}
\usage{
fileCountProject(projectAccession = c("PXD000001", "PXD000002"),
  accessionRownames = TRUE)
}
\arguments{
\item{projectAccession}{character, a vector of project accessions.
The default is c("PXD000001", "PXD000002")}
}
\value{
a data.frame, the number of files in the projects.
}
\description{
The number of files in the projects
}
\examples{
{
\dontrun{
fileCountProject("PXD000002")
fileCountProject(c("PXD000001", "PXD000002"))
}
}
}
