% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileListAssay.R
\name{fileListAssay}
\alias{fileListAssay}
\title{List files for a project}
\usage{
fileListAssay(assayAccession = c("22134", "22135"))
}
\arguments{
\item{assayAccession}{character, a vector of project accessions.
The default is c("PXD000001", "PXD000002")}
}
\value{
a data.frame of the details of searched entries, including:
projectAccession (string, optional): ,
assayAccession (string, optional): ,
fileType (string, optional) = ['RESULT' or 'PEAK' or 'SEARCH' or 'RAW' or 'QUANT' or 'GEL' or 'FASTA' or 'SPECTRUM_LIBRARY' or 'MS_IMAGE_DATA' or 'OPTICAL_IMAGE' or 'OTHER']: ,
fileSource (string, optional): SUBMITTED (part of the original dataset) or GENERATED (added to the submission by PRIDE),
fileSize (integer, optional): size in bytes,
fileName (string, optional): the name of the file,
downloadLink (string, optional): public FTP download link,
asperaDownloadLink (string, optional): public Aspera download link,
fileExtension: file extension,
fileSizeGB: file size in GB.
}
\description{
List files for a project
}
\examples{
{
\dontrun{
# Search for one accession
fileListAssay("22134")

# Search for two accessions
fileListAssay(c("22134", "22135"))
}
}
}
