% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prideDownload.R
\name{prideDownload}
\alias{prideDownload}
\title{The details of projects returned by searching}
\usage{
prideDownload(prideLinks, fileMaxSize = 10, fileType = "SEARCH",
  fileExtensionInclude = NULL, fileExtensionExclude = NULL,
  path = getwd())
}
\arguments{
\item{prideLinks}{data.frame, the result from \code{\link{fileListProject}}
or \code{\link{fileListAssay}} function.}

\item{fileMaxSize}{numeric, maximumal total file size.}

\item{fileType}{character, "RESULT", "PEAK", "SEARCH", "RAW", "OTHER".
The default is "SEARCH".}

\item{fileExtensionInclude}{character, the files with these extensions
are included to be downloaded. For example: "csv", "xlsx", "xls", "7z", etc.}

\item{fileExtensionExclude}{character, the files with these extensions
are excluded. For example: "raw", "fasta", "gz", "xml", etc.}

\item{path}{character, the path to save downloaded files.}
}
\value{
data.frame, information of downloaded files.
}
\description{
The details of projects returned by searching
}
\examples{
{
\dontrun{
# Search for one accession
fileInfo = fileListProject(c("PXD000001", "PXD000002"))
prideDownload(prideLinks = fileInfo,
              fileMaxSize = 10,
}
}
}
