% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectList.R
\name{projectList}
\alias{projectList}
\title{The project entries information returned by searching}
\usage{
projectList(query = "*", show = 10000, page = 0, sort = NULL,
  order = "asc", speciesFilter = NULL, ptmsFilter = NULL,
  tissueFilter = NULL, diseaseFilter = NULL, titleFilter = NULL,
  instrumentFilter = NULL, experimentTypeFilter = NULL,
  quantificationFilter = NULL, projectTagFilter = NULL)
}
\arguments{
\item{query}{character, the keywords to search. For example: "stress".}

\item{show}{integer, how many results to return per page. Maximum
page size is: 10000 (default).}

\item{page}{character, which page (starting from 0) of the result to
return. The default is 0.}

\item{sort}{character, the field to sort on (e.g. score, publication_date,
id or project_title).}

\item{order}{character, the sorting order ("asc" or "desc"). The default
is "asc"}

\item{speciesFilter}{numeric, filter by species. The NCBI taxon ID
(https://www.ensembl.org/info/about/species.html). For example,
9606 represents Homo sapiens (human), 10116 represents Rattus norvegicus
(rat), 9986 represents Oryctolagus cuniculus (rabbit), 10090 is Mus
musculus (mouse), and 7955 is Danio rerio (Zebrafish).}

\item{ptmsFilter}{character, filter by PTM (Post-translational modification)
annotation. For example: "phosphorylation".}

\item{tissueFilter}{character, filter by tissue annotation. For example:
"HEK-293 cell", "breast", "cell culture", "brain".}

\item{diseaseFilter}{character, the disease to filter. For example:
"disease free", "cancer".}

\item{titleFilter}{character, filter the title for keywords (example: stress).}

\item{instrumentFilter}{character, the instrument to filter. For example:
"LTQ Orbitrap", "LTQ", "LTQ FT", "LTQ Orbitrap Velos".}

\item{experimentTypeFilter}{character, filter by experiment type. For
example: "Bottom-up proteomics", "Affinity purification coupled with mass
spectrometry proteomics", "Shotgun proteomics", "shotgun".}

\item{quantificationFilter}{character, the quantification to filter. For
example: "Spectrum count/molecular weight", "Spectrum counting", "label-free".}

\item{projectTagFilter}{character, filter by project tags (example: Biomedical).}
}
\value{
a data.frame of the details of searched entries, including:
projectAccession, title, projectDescription, publicationDate, submissionType,
numAssays, species tissues, ptmNames, instrumentNames, and projectTags.
}
\description{
The project entries information returned by searching
}
\examples{
{
\dontrun{
# Search for stress
projectList(query = "stress")

# Search for stress and filtered by human
projectList(query = "stress", speciesFilter = 9606)

# Search for stress and filtered by human and sorted by publication_date
projectList(query = "stress", speciesFilter = 9606,
            sort = "publication_date", order = "desc")
}
}
}
